+++ os/audit_priv.h
